+++ packages/libdxfrw-converter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mlightcad/libdxfrw-converter
 
+## 3.0.4
+
+### Patch Changes
+
+- feat: enable web worker for dxf parser
+- Updated dependencies
+  - @mlightcad/data-model@1.2.4
+
 ## 3.0.3
 
 ### Patch Changes